Refactor: moved tipmanager to engine
