+++ pkg/protocol/engine/tipmanager/tip_metadata.go
